+++ src/main/java/org/openlmis/fulfillment/web/util/OrderLineItemDto.java
@@ -16,12 +16,17 @@
 package org.openlmis.fulfillment.web.util;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.apache.commons.collections.map.HashedMap;
 import org.openlmis.fulfillment.domain.OrderLineItem;
 import org.openlmis.fulfillment.service.ExporterBuilder;
 import org.openlmis.fulfillment.service.referencedata.OrderableDto;
@@ -31,10 +36,13 @@ import org.openlmis.fulfillment.service.referencedata.OrderableDto;
 @AllArgsConstructor
 public class OrderLineItemDto implements OrderLineItem.Importer, OrderLineItem.Exporter {
 
+  public static final String PRICE = "price";
   private UUID id;
   private OrderableDto orderable;
   private Long orderedQuantity;
   private Long totalDispensingUnits;
+  @Setter
+  private Map<String, Object> extraData;
 
   /**
    * Create new instance of TemplateParameterDto based on given {@link OrderLineItem}.
@@ -56,5 +64,18 @@ public class OrderLineItemDto implements OrderLineItem.Importer, OrderLineItem.E
         .ofNullable(orderable)
         .map(item -> new VersionIdentityDto(orderable.getId(), orderable.getVersionNumber()))
         .orElse(null);
+  }
+
+  @JsonIgnore
+  @Override
+  public Map<String,Object> getExtraData() {
+    Map<String,Object> extraData = new HashedMap();
+    BigDecimal price = BigDecimal.ZERO.stripTrailingZeros();
+    if (this.getOrderable() != null
+            && this.getOrderable().getPrograms() != null) {
+      price = this.getOrderable().getPrograms().iterator().next().getPricePerPack();
+    }
+    extraData.put(PRICE,price);
+    return extraData;
   }
 }