+++ src/main/resources/db/migration/20210305105532947__alter_table_orders.sql
@@ -1,0 +1,1 @@
+ALTER TABLE fulfillment.orders ADD COLUMN extradata jsonb;